requirements.txt
python-telegram-bot
PIL
selenium
validators
sqlalchemy